[core] Move interfaces of UI views to dedicated files
