+++ packages/x-date-pickers/src/DateTimePicker/shared.ts
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { useThemeProps } from '@mui/material/styles';
 import { useDefaultDates, useUtils } from '../internals/hooks/useUtils';
-import { ExportedTimeClockProps } from '../TimeClock/TimeClock';
-import { ExportedDateCalendarProps } from '../DateCalendar/DateCalendar';
+import { ExportedTimeClockProps } from '../TimeClock/TimeClock.types';
+import { ExportedDateCalendarProps } from '../DateCalendar/DateCalendar.types';
 import { DateTimeValidationError } from '../internals/hooks/validation/useDateTimeValidation';
 import { ValidationCommonProps } from '../internals/hooks/validation/useValidation';
 import { BasePickerProps } from '../internals/models/props/basePickerProps';
