+++ packages/x-date-pickers/src/YearCalendar/index.ts
@@ -1,5 +1,5 @@
 export { YearCalendar } from './YearCalendar';
-export type { YearCalendarProps } from './YearCalendar';
+export type { YearCalendarProps } from './YearCalendar.types';
 
 export { yearCalendarClasses, getYearCalendarUtilityClass } from './yearCalendarClasses';
 export type { YearCalendarClasses, YearCalendarClassKey } from './yearCalendarClasses';
